refactor: CertItem 리팩터링
