+++ frontend/src/components/CertItem/type.ts
@@ -1,5 +1,3 @@
 import { Cycle } from 'commonType';
 
-export type CertItemProps = Cycle;
-
-export type useCertItemProps = Pick<CertItemProps, 'startTime' | 'progressCount'>;
+export type CertItemProps = Omit<Cycle, 'challengerCount'>;
